requirements.txt
asgiref==3.8.1
autoflake==2.3.1
autopep8==2.3.2
Booktype==1.5
channels==4.2.2
Django==5.1.7
djangorestframework==3.16.0
flake8==7.1.2
iniconfig==2.0.0
isort==6.0.1
mccabe==0.7.0
packaging==24.2
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
pygame==2.6.1
pytest==8.3.5
redis==5.2.1
setuptools==76.0.0
simplejson==3.20.1
sqlparse==0.5.3
